Fix bug in localised articles
When building the legacy tagging page, I should have use the
LocalisedModel class used by the original editions controller.

This introduced a problem where the edition was not localised
correctly.

It also broke the "Create a News article of type 'world news story'
in a non-English language" scenarion in news-article.feature