Change button text on Edit page
Change button text from "Save" and "Next" to "Save" and
"Save and continue".

Whitehall uses the same buttons on every type of Edit page,
so changes don't happen in the view as expected, but in a
form builder that is used in every "form" partial for every
type of edition.

The form builder already had a function for "save and continue" so
one didn't need to be created.

Update feature tests to change text of button clicked.