Make the "Save topic changes" button green
The button colour was being set by `form_actions` in the 
`Whitehall::FormBuilder` class. However, that method renders
all buttons passed to it in blue.

This can probably be refactored to update `form_actions` to get it
to accept a "colour" value being passed in, and leaving the default 
as blue, but for expediency, the div that `form_actions` would create
has been hard-coded into the view.