Remove "required" asterisk for policy areas
If a piece of content can be tagged to the topic taxonomy it
is no longer required for them to also be tagged to a policy area.

Seeing a red asterisk on the legacy tagging page could mislead
a user into extra tagging.

Created a new function to determine whether it's mandatory to
tag a policy area as the partial is also used for world location news
content and speed tagging.