Fix the delay of topic taxons appearing on the editions overview page
Currently, when rendering the overview edit page, after selecting the
Topic Taxons you want to tag the edition too, the taxons sometimes don't
show in the topic taxon display box.

This is due to the request getting the Publishing API's stored
data which might be out of date, adding `generate: true` into the
request requires publishing-api to generate the expanded links instead
of relying on its already stored expanded links. This should ensure that
the most up to date links are returned. Meaning there now should be no
delay in the displaying of taxon information.