Remove some inconsistency in the taxonomy tree data
Previously the root hash would have a `links` value, and also a
`expanded_links_hash` value. This is because of the way the expanded
links payload for the homepage is split up, and spliced together with
the expanded links payload for each level one taxon.

All JSON payloads include the "expanded" links from the Publishing
API, and the name "links" is used throughout the payloads for
this. This change inserts the expanded links for each level one taxon
in to this place in the hash for the level one taxon, allowing the
code to handle the data in a more uniform manor.

I was looking in to this due to the mapping code not handling the
legacy_taxons for level one taxons correctly, which was due to the
code previously not knowing to look for the data in two places due to
this edge case. This is now unnecessary, and the code works.