Fix the Worldwide tagging form when no tags are selected
Previously it would error in the controller as nothing was in the
params with the name `taxonomy_tag_form`. The functional part of the
form is in a partial as it's currently shared between the Worldwide
and Taxonomy tagging interfaces, and this works fine for the Taxonomy
Tagging form, just not the Worldwide one.

It works for the Taxonomy tagging interface, as in that case, there
are other things in the `taxonomy_tag_form` entry in params, so to
keep the code somewhat consistent, change the name of the value in
params.

In the future, these interfaces will hopefully diverge further, as the
Topic Taxonomy tagging interface doesn't suit the data in the
Worldwide tagging interface. At which point this change will hopefully
no longer be needed/relevant.