Fix future test dates
Many Whitehall tests are frozen by Timecop at 11:11 11 Nov 2011
which means that validating the first_published_at date cannot
be in the future makes these particular tests fail as they use
dates since Nov 2011. I've adjusted the dates but ideally we should
revisit these and other related tests so that they don't rely on
a fixed date in the setup.