Track taxon selections on either save button
Send an event to GA with the selections the user has made, ideally once
(on save) with the complete list of selected taxons.

This commit sends the GA event once for each selected taxon.

This happens when the use presses the 'Save topic changes' or
'Save and review legacy tagging' buttons.