Make Worldwide taxons expand/collapse per country
Currently, only `world` expands which displays a huge list of all
the countries and their sub taxons. This is not good for the user as its
an overload of information.

This change provides a link to expand/collapse per country, this new list
also displays everything alphabetically starting with all the taxons
that have no children.