Separate DocumentListExportWorker to own queue
I would like to separate this worker, to have it's own queue, as I
believe the emails are not being sent in production.

We currently have a few errors regarding the size of the attachment
being superior to 10MB (https://sentry.io/govuk/app-whitehall/?query=is%3Aunresolved+Net%3A%3ASMTPFatalError).

I want to narrow down if the queue is too big in Production that the
emails are not being triggered.