Limit the number of records which can be exported.
Large resultsets of unfiltered documents can total around 250,000
records which makes for a > 10MB email attachment, which then fails
to be sent because of notification limits.
It also loads the backend quite heavily so adding a restriction with
an appropriate message is the agreed protection against this.
The limit is 8000 which after some testing appears to be just below
the 10MB attachment threshold.