Render org pages from collections
When we've done the visual regression testing and are happy, we should merge
this so that updates to org pages will make them render from collections.

Republishing orgs takes all day at the moment, so we'll add a task to allow us to
switch all routes quickly without a republish.  We can also switch back if it
doesn't go so well.