Add field for important board members
This commit adds a new field to the organisation presenter to denote the number of important board members, which are board members who have their photos displayed on the organisation home page. It also removes the former logic which only presented photos for such people since it didn’t work for people with multiple roles.