Revert "Fix the delay of topic taxons appearing on the editions overview page"
Reverts alphagov/whitehall#4094

We need to revert to using the cached endpoint for expanded links. Constantly hitting the uncached endpoint is causing very high load on Postgres, and giving users 500s. We think this bug existing may be preferable to serving lots of errors and reducing platform stability.

This is a temporary revert while we investigate a longer term solution.