Set uploaded_to_asset_manager timestamp in a worker
There exist a few `AttachmentData`s which do not have an
`uploaded_to_asset_manager_at` timestamp, and which don`t correspond
to any `AssetManagerCreateWhitehallAssetWorker` exceptions in the
logs.  The worker succeeded, but failed to set the timestamp.  This
means that the `AttachmentData` for the given `legacy_url_path` was
not found.

Having looked at a few of these assets, the `AttachmentData` in
Whitehall and the `WhitehallAsset` in Asset Manager were created in
the very same second.

Given that there doesn`t appear to be any direct causal link between
creating the `AttachmentData` and running the worker, I suspect there
is a race condition between the two: if the worker runs to completion
before the `AttachmentData` is created (which it could feasibly do for
small files when the Sidekiq queue is otherwise empty), then the
`AttachmentData` will not exist when the worker tries to set its
`uploaded_to_asset_manager_at`.

This commit splits out the setting of the timestamp to a separate
worker, which throws an exception (and so gets retried) if it fails to
find the `AttachmentData`.  This is a bit of noise, which is sad, but
it does only appear to affect a very small proportion of uploads.