Update asset redirect url on unpublish
AttachmentUpdater will update all asset fields on most
Edition workflow operations. This makes it difficult
to clear the redirect_url for attachments which have been
added to a draft which has been previously unpublished.
This prevents editors from previewing the newly added attachment.

This commits changes this behaviour so that the redirect_url
is only written on unpublish and can be updated to nil by
subsequent operations. This allows an attachment on a (re)drafted
edition to be previewable.