Ignore AttachmentDataNotFound exception
This exception was introduced because sometimes the AttachmentData
doesn't exist by the time the asset is uploaded, so we need to try
again after a little while.

It's a suboptimal solution, but it seems to work.  Ideally there would
be an ordering constraint between creating the AttachmentData and
running the worker, but I'm not sure that's feasible right now.