Recognise PDF thumbnail paths in SetUploadedToWorker
The legacy URL path of a PDF thumbnail won't match the legacy URL path
of the PDF it belongs to, so workers created for thumbnail uploads
were failing.

We don't call uploaded_to_asset_manager! when we upload a thumbnail,
because the actual file may not have made it across yet.  In contrast,
if the PDF makes it across but not the thumbnail, it's not such a big
deal.