Export news documents as JSON
This is to provide the functionality to export the data for news
documents so that they can be imported into a different publishing
application. They are exported in a JSON format with most of the known
data associated within a news document/edition with the exception of
Versions.

The usage is:

rake export:news_documents ORGS=org-slug FROM=2018-01-01 OUTPUT=/path/to/file"

where
ORGS is a comma-separated list of organisation slugs (e.g.
hm-revenue-customs,ministry-of-defence) and is a required option
FROM is a parsable date string that MySQL can understand
OUTPUT is a path to the file to write the output too

This task is focussed on news as that is the first format we are
considering for moving to a new publishing application. This could be
developed further to cover all document types, however we would need to
learn further about their individual fields and associations.