Apply `-dUseCropBox` when generating PDF thumbnails
The issue:
Certain PDF thumbnails contain extra blank space.

The cause:
The MediaBox and CropBox settings of the first page of those PDFs do not match, and the PDFFitPage option is not cropping.

The solution:
Use the CropBox setting to properly crop the page while keeping PDFFitPage to keep the image scaled down.

References:
https://stackoverflow.com/questions/27779246/when-converting-first-page-of-a-pdf-into-an-image-using-ghostscript-sometimes-i