Support running the sanitise db test with a custom host and port
Without this change, the default host and port are used. Adding the -H
option, and setting the MYSQL_TCP_PORT environment variable means that
the ActiveRecord host and port settings are respected.

This makes it possible to run this test when not using the standard
host and port for MySQL.