Add the rails_server_timings gem
This exposes the timings information (view, database) reported to
ActionController as headers in the HTTP responses. This makes it
possible to view how long the server spent on processing different
parts of the response, right in the browser you're viewing it with.

Disable this functionality in production, as it's just for local
development.