Track and log Publishing API Subscriber runtime information
ActiveController logs runtime information, and other components like
ActiveRecord feed in to this. This is exposed through the logs, and
also via the just added rails_server_timings gem.

This commit wraps the PublishingApiPusher service listener in a way
that the timing information is recorded, and then reported both in the
logs (example below), and as part of the Server-Timings header.

If this works in Whitehall, it might be worth trying to extract this
out in to the gds-api-adapters gem.

```
    ...
    Publishing API (166.4ms) [ update_draft, options: {} ]
    ...
  Completed 302 Found in 268ms (ActiveRecord: 21.3ms | Publishing API: 166.4ms)
```