Render announcements on people page using rummager
This converts the "Announcements" section on people pages to use the
search API. This means that content published by other applications
(like upcoming content-publisher) will be surfaced on this page too.

The short/medium term plan for these pages is that they're moved to the
collections app. To make this easier, I've chosen to use some
components to rebuild the interface. This way I can avoid refactoring
whitehall code to make use of the search data structure instead of the
data from the database (the `feeds` partial is tightly coupled to the
active record models). It hopefully makes the interface better as well.

https://trello.com/c/vYdbh3T