Add task to remove policy area
This rake task unpublishes a policy area and sets up a redirect
for the given `base_path`.

Index is removed from rummager and the state of the
policy area in whitehall is set to `deleted`.

Functionality has also been removed that stops policy areas being
removed if they have accosiated policies. This isn't that relevent
anymore as policies are soon to be retired and the only place this is
used is in the tagging interface.

Trello;
https://trello.com/c/HbFEVJX3/135-l-add-rake-task-to-whitehall-to-unpublish-policy-areas