Skip Publishing API callbacks when seeding the database
As the Publishing API might not be available.

This change is necessary, as now the Publishing API jobs are not
delegated to Sidekiq, but processed immediately, and the Publishing
API might not be available when seeding the database.

Topic's, which are also called Policy Areas are being retired soon,
which should remove the need for this change. As for Organisations,
there's also a plan to move the handling of them to another app, which
hopefully won't store local state that it needs to sync with the
Publishing API (therefore avoiding this problem).