Remove references to Policies and Policy Areas
When a Whitehall user creates a new document, we allow them to associate
it with the legacy associations. As we are in the process of
unpublishing/retiring all Policies and Policy Areas, we should remove
this possibility, once they are not present.

With this, we now show only the legacy associations that are still
present.

Once all the Policies and Policy Areas are unpublished, we can modify
this and remove the conditionals introduced.