Don't send the update_type when publishing
The Publishing API can accept the update_type when publishing an
edition, but this isn't ideal. The update type should be provided when
creating the draft (with the put content request), this ensures things
are previewed accurately for example.

Therefore, switch to not providing the update_type when
publishing. Currently, the GDS API Adapters, want to send null for
this, so I've been replacing it with nil. Whitehall still sends a put
content request prior to publishing, so this shouldn't affect the
actual update type used.