Support subtopic filtering in the document filter
This is primarily to expand support for more complicated email
subscriptions to the Topic Taxonomy, beyond the level one topics. This
in turn, should provide a better experience for those who have
subscriptions to Policy Areas which are being ended, as most Policy
Areas are similar to level two Topics in the Topic Taxonomy.

Unfortunately, due to time constraints, the interface for this is a
bit crude, involving a second select input, populated with the second
level Topics. This looks good to me in the interface, but in the code,
it's frustrating, as this overlaps with the existing Topic select
input.

There also isn't such a thing as a subtopic, it's the term used in the
interface, and for consistency it's used in the code as well, but in
this context, it just means level two Topic in the Topic
Taxonomy. Hopefully if a better interface can be built for this, the
code and terminology can be improved.

The second level Topics are fetched using the Taxonomy::TopicTaxonomy
class. This is in contrast to the Content Store approach used for the
level one Topics, as to enable the JavaScript powered automatic
refresh behaviour, all the second level Topics need to be represented
on the page, even if only the relevant ones are shown to the
user. Fetching all of the second level Topics from the Content Store
would currently require a request per level one Topic, so using the
data cached in Redis should be faster.