Remove LevelOneTaxonsFetcher from DocumentFilter::Options
Replace it with the Taxonomy::TopicTaxonomy class, which fetches the
data from the Publishing API, rather than the Content Store. This
should be quicker than fetching the data for the content store, and is
more consistent with the fetching of the level two topics, which uses
the Taxonomy::TopicTaxonomy route.