Don't reuse content ID for new HTML attachments
This was put in with an understanding that the publishing API would reject new HTML attachments on the same base path if a previously deleted HTML attachment had left a redirect there.

This doesn't make sense because the publishing API is a) supposed to allow the overwriting of redirects in all cases and b) is supposed to allow the overwriting of any content item if the publishing application matches.

This will be checked before deploy, but shouldn't be necessary.