Display full summary for classification featurings
When entering a summary for a document in Whitehall, guidance encourage users
submit 160 characters or fewer. However when a document is featured on a
topical events page, the summary is truncated to 150 characters. This changes
the limit to 160 characters - it is still possible for users to enter more than
this limit, so it still needs to be truncated for these cases.