Separate updating an edition from publishing
Previously, the publish action in Whitehall was bound to an update of
the edition. This is wasteful, as the content in the Publishing API
should be what is being published already, and risky, as updating it
could mean that something different from what the user has previously
previewed is published.