Untangle patching links from publishing
Previously, the links would be patched when publishing. As Whitehall
mostly uses Edition links, I'm not sure if this actually affects
anything, but to make the publishing action neater, and avoid problems
with link updating conflicts caused by this extra update, separate it
out from the standard publish method in the PublishingApi class.