Remove the 'Coming soon' related functionality
Working from the description in the ComingSoonPresenter, this was a
workaround for issues with caching in the wider publishing
platform. Now, the publishing platform handles cache invalidation, so
this shouldn't be a problem any more.

The functionality in Whitehall sits both around the scheduling, and
unscheduling, so I've removed the coming soon part of both actions.