Patch links when updating drafts
Whitehall seems split between using ediiton links, and non-edition
links for different document types. Previously, a patch_links request
was made when publishing content, which is not the ideal time (for
reasons like draft preview inconsistencies).

So, patch the links of content when saving drafts, as this will make
changes like tagging to organisations take effect.