Separate out the edition links in the presenters
For the Publishing API. Some presenters put the links in to the
edition, but these links will also be sent to the Publishing API using
a patch_links request.

So, to unmuddle this, rename links in the relevant presenters to
edition_links, and add a new links method which returns for now
returns the same value. Hopefully the two can then be tweased apart,
bit by bit, reducing the duplicate information being sent to the
Publishing API.