Avoid issues with saving localised content
Previously, the LocalisedModel created in the controller was being
passed through to the Whitehall::PublishingApi class. The
LocalisedModel doesn't work the same as a normal model though, and it
breaks the way the PublishingApi class sends the translations to the
Publishing API.

Calling reload seems to avoid this broken behaviour. Unfortunately, I
don't know why LocalisedModel is used, so I'm unsure if this is the
best way of fixing this, but it seems to work.