Revert "Separate updating an edition from publishing"
What should have been a simple code cleanup has caused multiple
issues, including problems with scheduled publishing, timestamps and
change notes, simply due to Whitehall never being changed to send
drafts to the Publishing API which are actually suitable for
publishing.

Some of these issues have been resolved, like the scheduled publishing
and timestamp issues, but at this point, for the change notes, the
relevant code doesn't even attempt to create something suitable, and
it doesn't look like a simple change, as with the way Whitehall
"republishes" content, if you were going to change it to include
drafts, this would have to be conditional on whether you plan to
immediately publish the content, or to leave it as an actual draft.

So, this commit returns to not publishing the draft in the Publishing
API, but instead sending through a new draft after the state changes
in Whitehall have taken place, which is then immediately
published. This isn't ideal, as obviously it defeats the purpose of
content preview if you're not publishing what you've previewed, but
this isn't the time to attempt to fix that.

This reverts commit 6ea0aca7e55bfd5a2451ca96928252de41007c6c.