Ensure current_release_date is up-to-date
Due to how associations work in ActiveRecord, I think there could be
issues with changing dates for StatisticsAnnouncements, as the
association on the model doesn't know that the data it has is now
invalid if the say a new StatisticsAnnouncementDate is created.

Therefore, to avoid this, reload the association after a
StatisticsAnnouncementDate is saved.

I think the test changes relate to this change having effects on the
test setup, and not actually a sign that the tests were wrong (well, a
bit, but not in a meaningful way).