Don't try to reslug when > 1 document has a slug
Slugs are not unique to Documents. We can have documents of different
formats sharing slugs. So we can have a Policy Paper have a slug
"my-slug" and also a Speech have a slug "my-slug".

Ensure we warn anyone running the document resulgging rake task that
there are multiple documents with the same slug so they do not reslug
the wrong document by mistake.