Add rake task rummager:index:topical_event_editions
This rake task will enable re-indexing of documents related to topical events.

This is useful if there is a change to topical events, and we need to re-index
all related documents.

For the completion of https://github.com/alphagov/rummager/pull/1298 this
rake task will need to be run on all environments.

The rake task intends to get all documents related to a topical event,
and that belong in rummager (published editions), and then add these
documents to the queue for re-indexing in rummager.

In addition, this commit adds topical_events to documents sent to rummager,
in place of policy_areas, which are due for removal.