Make scheduled publishing more reliable
For the majority of content for which being published by the
Publishing API is required for it to be "published", Whitehall still
doesn't consider the Publishing API interaction to be a critical
component.

This is a general problem, but specifically for scheduled publishing,
this is causing an issue because it can be a cause of extra load on
the Publishing API, to the degree that it causes timeouts. When this
happens, the ScheduledPublishingWorker jobs fail, but when they're
retried, they no-op, as the Edition has already been transitioned to
the published state.

Extending the transaction around the whole operation of the
ScheduledEditionPublisher service in this case, means that the
Publishing API interaction (along with all the other edition services
listening in) is part of the transaction, and thus if it fails, the
database won't be updated to say that the edition has been
successfully published.

This change could be made in the EditionService class, but as the
current problems are focused around scheduled publishing, making such
a wide reaching change seems too risky without enough benefit, when
making it here should have the same effect, but just for scheduled
publishing.