Handle the Publishing API within the EditionService
Rather than processing the Publishing API change after the change has
been saved in Whitehall, process it during the transaction. This means
that there is less opportunities for Whitehall's state to diverge from
the Publishing API, as the state will only be transitioned in
Whitehall when the Publishing API update succeeds.