Use Rummager for related document lists on topical events
This change introduces RummagerDocumentPresenter, a wrapper for documents
returned by Rummager searches. It changes the source for related documents
on topical event pages from Whitehall to Rummager. For each related document
list on the topical event page, we make a rummager search query.

In addition it introduces additional views and test logic. This is necessary
because other pages will still have lists with documents sourced from Whitehall.

Hopefully RummagerDocumentPresenter will be extended and used more as we transition
away from Whitehall to using content store, publishing api and rummager to back
Finders and other kinds of list.

A side effect of this change is that the atom feed results for the topical events
will change. This may result in feed readers interpreting old documents as new,
since the IDs and other attributes will appear to have changed.