Short circuit has_links for editions that are the length of a novel
We have an edition that is over 200000 words long. That's longer than
many novels.

This is causing timeouts when `LinkCheckerApiService#has_links?` tries
to extract the links from the body of the edition.
`LinkCheckerApiService#has_links?` is called indirectly from
`editions/show.html.erb`. The results are used to decide whether or
not to show the link checking sidebar (i.e. when no links are present
in the edition's body, we don't bother to show the link check
sidebar).

This PR short circuits the logic for editions that are over 99999
words long and just displays the link check sidebar anyway. This does
potentially introduce an inconsistency where a very long edition
without links will show the link check sidebar regardless. However,
this does also get around our timeout for now and will allow editors
to actually get to their editions.

I've chosen 99999 words because
[Wikipedia](https://en.wikipedia.org/wiki/Word_count#In_fiction)
suggest a novel to be 100000 words.