Less empty data in Contacts presented to Publishing API
Many of contacts in Publishing API have big chunks of empty data in them
which causes poor rendering.

Example:

```
{:description=>"Email monitored Monday - Friday, 9.00am - 5.00pm",
 :contact_type=>"Media contact",
 :title=>"Press Office (media enquiries)",
 :contact_form_links=>
  [{:title=>"Press Office (media enquiries)", :link=>"", :description=>""}],
 :post_addresses=>
  [{:title=>"",
    :street_address=>"",
    :locality=>"",
    :region=>"",
    :postal_code=>"",
    :world_location=>""}],
 :email_addresses=>[{:title=>"", :email=>"comms@nio.gov.uk  "}],
 :phone_numbers=>
  [{:title=>"Belfast", :number=>"028 9052 7856 "},
   {:title=>"Out of hours duty press officer",
    :number=>"M:  ‭‭+44 ‭+44 7464 981449‬"}]}
```

To resolve this this uses the presence methods on strings or nils to
make them nil and then uses compact to remove them from hashes.

This also removes the title from the contact form as it did not make
sense to duplicate the contact title here (and is not required by the
schema)