Export contact lookup tables with news document
This extends the ExportNewsDocument class to also include information of
the contacts which are embedded in a document.

When this data is then imported into Content Publisher we can then
switch around the contacts to be set via content_id rather than id.

Since this added a reasonable bit of logic I felt it was best to have a
test accompany it.