Add display_data_microformat to Rummager Document Presenter
The mustache template that renders the announcements search results
expects to be to able to call `display_data_microformat` and have it
return a block of html.

This method is called from a view helper. As we are returning this code
from a presenter instead, it wasn't enough to just include the module
that contains `content_tag`, we also need to inherit from ActionView to
handle the output in `output_buffer`.

See https://stackoverflow.com/questions/5452646/undefined-method-output-buffer-for-content-tag-with-block
and https://api.rubyonrails.org/classes/ActionView/Context.html for more
information.