Add an `as_hash` method to RummagerDocumentPresenter
This method was provided by the Filterer class for use in the mustasche
templates that render the announcements finder. As we displaying results
directly from rummager we need to provide an equivalent.

The announcements finder view expects to assign a unique id in the CSS for each
result, so we use each document's content_id to ensure uniqueness.