Replace SearchResult with RummagerDocumentPresenter
The AnnouncementPresenter is tightly coupled to the Announcement model.
As we are getting the search results from Rummager and not the database,
we can no longer use this presenter.

RummagerDocumentPresenter has been updated to provide what is needed in
the views. As this duplicates some of what SearchResult was already
doing, we have replaced that too.