Render latest page Rummager results
Now we have switched to use Rummager to retrieve documents for the latest
finder, we create a partial which is able to work with Rummager results.

As part of this work we investigated whether we needed to also retrieve
translated documents from the Whitehall database in cases where the locale is
not English (Rummager currently doesn't support locales). Our investigation
concluded that there are no apparent cases where the locale is set to a foreign
language; for example there don't seem to be any instances where a locale param
is passed to the Latest Controller. A manual check on live GOV.UK also
concluded that links from translated world location news pages (e.g.
http://www.gov.uk/world/afghanistan/news.dr) to the latest finder also don't
pass any locale data, and so the results on the page are English only.
Consequently we've decided to maintain this current behaviour and only provide
English-language Rummager results.