Topical Events announcements filtered by content_store_document_type
This switches the rummager search for announcements on Topical Events pages to
filter by `content_store_document_type` instead of `format`. We do this for a
few reasons:

1) The announcements finder now filters by `content_store_document_type` as it
allows more precise control over the documents returned. For consistency We
should ensure that the 3 documents returned in the announcements section on
topical events pages match the top 3 documents that would appear on an
announcements finder filtered by a given topical event.

2) The current filtering only filters for `news_article`, but it's possible
that other announcement types could be tagged to a topical event.

3) `format` is expected to be deprecated in Rummager

We also ensure that `content_store_document_type` field is returned from
Rummager, as this will enable displaying the document's type in the
announcements section for Content Publisher documents.