Add shared document_list_from_rummager partial
We duplicate the `document_list_from_rummager` partial that belongs to
classifications and make it available to be shared amongst multiple views. For
now we change the world location news page to use this partial to render its
announcements section with the view to replace usage of the old classifications
partial with the new one on topical events pages in a future commit.