Alter LatestDocumentFilter to filter rummager results
This class filters and paginates the results returned by
SearchRummagerService.

`Classification` has also been renamed to `TopicalEvent`, this is due to
classifications orginally meaning topical events and policy areas.
Policy areas have now been retired and referring to topical events as
classifications is vague.

Rejected types:

For orgs, documents which have the format of
`corporate_information_page` are rejected. This doesn't actually matter
as these pages are legacy and have been replaced by finder-frontend.
However the pages are still live if you type in a specific url.

Documents with the type of `world_location_news_article` are currently
rejected for latest documents relating to topical events, to replicate
this behaviour `news_article`'s are rejected from the returned documents
due to it being the only content_store_document_type which directly maps
to the format of `world_location_news_article`.

Format isn't used because it doesn't exist in the govuk index and is
only a thing in whitehalls index.